backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-docx==1.1.0
pdfplumber==0.10.3
pygments==2.16.1
playwright==1.40.0
docker==6.1.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27
pillow==10.1.0
aiofiles==23.2.1
jinja2==3.1.2
openai==0.28.1
jsonschema==4.19.0
